chore: refactor to simply graph check
Move util function out of graph/check.go into other files to keep
graph/check and its associated unit test readable.
